fix: recurcive call in foundation amino
