+++ packages/finschia/src/index.ts
@@ -49,7 +49,7 @@ export {
   setupCollectionExtension,
 } from "./modules";
 export { EvidenceExtension, setupEvidenceExtension } from "./modules";
-export { FeeGrantExtension, feegrantTypes, setupFeeGrantExtension } from "./modules";
+export { FeeGrantExtension, setupFeeGrantExtension } from "./modules";
 export {
   createFoundationAminoConverters,
   createMsgGrant,
@@ -95,7 +95,7 @@ export {
   setupFoundationExtension,
   ThresholdDecisionPolicyEncodeObject,
 } from "./modules";
-export { IbcExtension, ibcTypes, setupIbcExtension } from "./modules";
+export { IbcExtension, setupIbcExtension } from "./modules";
 export { stakingplusTypes } from "./modules";
 export {
   isMsgBurnEncodeObject,
@@ -132,9 +132,6 @@ export {
   wasmTypes,
 } from "./modules";
 export { makeLinkPath } from "./paths";
-export {
-  finschiaRegistryTypes,
-  SigningFinschiaClient,
-  UploadAndInstantiateResult,
-} from "./signingfinschiaclient";
+export { SigningFinschiaClient, UploadAndInstantiateResult } from "./signingfinschiaclient";
+export { finschiaRegistryTypes } from "./types";
 export { longify } from "./utils";
