--- packages/finschia/src/modules/feegrant/messages.ts
@@ -1,7 +1,0 @@
-import { GeneratedType } from "@cosmjs/proto-signing";
-import { MsgGrantAllowance, MsgRevokeAllowance } from "cosmjs-types/cosmos/feegrant/v1beta1/tx";
-
-export const feegrantTypes: ReadonlyArray<[string, GeneratedType]> = [
-  ["/cosmos.feegrant.v1beta1.MsgGrantAllowance", MsgGrantAllowance],
-  ["/cosmos.feegrant.v1beta1.MsgRevokeAllowance", MsgRevokeAllowance],
-];
