--- packages/finschia/src/modules/ibc/messages.ts
@@ -1,48 +1,0 @@
-import { GeneratedType } from "@cosmjs/proto-signing";
-import { MsgTransfer } from "cosmjs-types/ibc/applications/transfer/v1/tx";
-import {
-  MsgAcknowledgement,
-  MsgChannelCloseConfirm,
-  MsgChannelCloseInit,
-  MsgChannelOpenAck,
-  MsgChannelOpenConfirm,
-  MsgChannelOpenInit,
-  MsgChannelOpenTry,
-  MsgRecvPacket,
-  MsgTimeout,
-  MsgTimeoutOnClose,
-} from "cosmjs-types/ibc/core/channel/v1/tx";
-import {
-  MsgCreateClient,
-  MsgSubmitMisbehaviour,
-  MsgUpdateClient,
-  MsgUpgradeClient,
-} from "cosmjs-types/ibc/core/client/v1/tx";
-import {
-  MsgConnectionOpenAck,
-  MsgConnectionOpenConfirm,
-  MsgConnectionOpenInit,
-  MsgConnectionOpenTry,
-} from "cosmjs-types/ibc/core/connection/v1/tx";
-
-export const ibcTypes: ReadonlyArray<[string, GeneratedType]> = [
-  ["/ibc.applications.transfer.v1.MsgTransfer", MsgTransfer],
-  ["/ibc.core.channel.v1.MsgAcknowledgement", MsgAcknowledgement],
-  ["/ibc.core.channel.v1.MsgChannelCloseConfirm", MsgChannelCloseConfirm],
-  ["/ibc.core.channel.v1.MsgChannelCloseInit", MsgChannelCloseInit],
-  ["/ibc.core.channel.v1.MsgChannelOpenAck", MsgChannelOpenAck],
-  ["/ibc.core.channel.v1.MsgChannelOpenConfirm", MsgChannelOpenConfirm],
-  ["/ibc.core.channel.v1.MsgChannelOpenInit", MsgChannelOpenInit],
-  ["/ibc.core.channel.v1.MsgChannelOpenTry", MsgChannelOpenTry],
-  ["/ibc.core.channel.v1.MsgRecvPacket", MsgRecvPacket],
-  ["/ibc.core.channel.v1.MsgTimeout", MsgTimeout],
-  ["/ibc.core.channel.v1.MsgTimeoutOnClose", MsgTimeoutOnClose],
-  ["/ibc.core.client.v1.MsgCreateClient", MsgCreateClient],
-  ["/ibc.core.client.v1.MsgSubmitMisbehaviour", MsgSubmitMisbehaviour],
-  ["/ibc.core.client.v1.MsgUpdateClient", MsgUpdateClient],
-  ["/ibc.core.client.v1.MsgUpgradeClient", MsgUpgradeClient],
-  ["/ibc.core.connection.v1.MsgConnectionOpenAck", MsgConnectionOpenAck],
-  ["/ibc.core.connection.v1.MsgConnectionOpenConfirm", MsgConnectionOpenConfirm],
-  ["/ibc.core.connection.v1.MsgConnectionOpenInit", MsgConnectionOpenInit],
-  ["/ibc.core.connection.v1.MsgConnectionOpenTry", MsgConnectionOpenTry],
-];
